Move setup_logging under cli/
