+++ src/cloudai/__main__.py
@@ -14,55 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import logging
-import logging.config
-
-from cloudai.cli import CloudAICLI
-
-
-def setup_logging(log_file: str, log_level: str) -> None:
-    """
-    Configure logging for the application.
-
-    Args:
-        log_level (str): The logging level (e.g., DEBUG, INFO).
-        log_file (str): The name of the log file.
-    """
-    numeric_level = getattr(logging, log_level.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError(f"Invalid log level: {log_level}")
-
-    LOGGING_CONFIG = {
-        "version": 1,
-        "disable_existing_loggers": True,
-        "formatters": {
-            "standard": {"format": "%(asctime)s - %(levelname)s - %(message)s"},
-            "short": {"format": "[%(levelname)s] %(message)s"},
-        },
-        "handlers": {
-            "default": {
-                "level": log_level.upper(),
-                "formatter": "short",
-                "class": "logging.StreamHandler",
-                "stream": "ext://sys.stdout",
-            },
-            "debug_file": {
-                "level": "DEBUG",
-                "formatter": "standard",
-                "class": "logging.FileHandler",
-                "filename": log_file,
-                "mode": "w",
-            },
-        },
-        "loggers": {
-            "": {
-                "handlers": ["default", "debug_file"],
-                "level": "DEBUG",
-                "propagate": False,
-            },
-        },
-    }
-    logging.config.dictConfig(LOGGING_CONFIG)
+from cloudai.cli import CloudAICLI, setup_logging
 
 
 def main() -> None:
